Make it possible to store references from one identifier to another
This commit adds methods to store and purge references from the identity
store. It also changes the sweep algorithm so that it will traverse such
references when it performs a sweep for a specific identity prefix. The
new functionality is needed when one workflow step calls on another
workflow (or explicit resource) because that workflow must be considered
part of subsequent garbage collections of the original workflow.

Relates to lyraproj/lyra#314